feat: adds types + updates to latest go parser, fixes #6
