+++ .eslintignore
@@ -1,3 +1,3 @@
 node_modules
 wasm_exec.js
-main.d.ts
+types.d.ts
