+++ go.mod
@@ -2,4 +2,4 @@ module github.com/tablelandnetwork/wasm-sqlparser
 
 go 1.16
 
-require github.com/tablelandnetwork/sqlparser v0.0.0-20220725134734-b79ca3e804c8
+require github.com/tablelandnetwork/sqlparser v0.0.0-20220830151425-aec4df78e4c8
