+++ package.json
@@ -4,20 +4,20 @@
   "files": [
     "main.wasm",
     "main.js",
-    "node.js",
-    "main.d.ts",
+    "module.js",
+    "types.d.ts",
     "wasm_exec.js"
   ],
   "exports": {
-    "node": "./node.js",
-    "import": "./main.js",
-    "require": "./node.js",
-    "default": "./main.js"
+    "node": "./main.js",
+    "import": "./module.js",
+    "require": "./main.js",
+    "default": "./module.js"
   },
   "type": "module",
-  "main": "node.js",
-  "module": "main.js",
-  "types": "main.d.ts",
+  "main": "main.js",
+  "module": "module.js",
+  "types": "types.d.ts",
   "devDependencies": {
     "assert": "^2.0.0",
     "eslint": "^8.22.0",
@@ -28,10 +28,13 @@
     "eslint-plugin-promise": "^6.0.0",
     "mocha": "^10.0.0",
     "prettier": "^2.7.1",
+    "tsd": "^0.24.1",
     "wasm-opt": "^1.3.0"
   },
   "scripts": {
-    "test": "npx mocha tests/*.js",
+    "test": "npm run test:types && npm run test:code",
+    "test:types": "npx tsd",
+    "test:code": "npx mocha tests/*.js",
     "lint": "eslint '**/*.{js,ts}'",
     "lint:fix": "npm run lint -- --fix",
     "prettier": "prettier '**/*.{ts,json,sol,md}' --check",
@@ -41,5 +44,15 @@
     "wasm:opt": "npx wasm-opt -O main.wasm -o main.wasm",
     "build": "npm run wasm:go && npm run wasm:opt",
     "clean": "rm -rf main.wasm"
+  },
+  "tsd": {
+    "directory": "tests",
+    "compilerOptions": {
+      "paths": {
+        "@tableland/sqlparser": [
+          "."
+        ]
+      }
+    }
   }
 }
