+++ types.d.ts
@@ -1,0 +1,44 @@
+// Type definitions for @tableland/sqlparser
+// Project: @tableland/sqlparser
+// Definitions by: Carson Farmer <carson@textile.io>
+
+declare module "@tableland/sqlparser" {
+  /**
+   * Initialize the @tableland/sqlparser module.
+   * This must be called first to populate the sqlparser global namespace.
+   *
+   * @param input Input is provided primarily for testing purposes. It can be an optional `string`,
+   * `Request`, or `URL` specifing a "remote" WASM source, or a `BufferSource` to specify a local WASM file.
+   * It is best to leave undefined, and allow the library to use the included local WASM binary.
+   * @return A `Promise` that resolves to a WASM `Exports` object.
+   */
+  export function init(
+    input?: string | Request | URL | BufferSource
+  ): Promise<WebAssembly.Exports>;
+
+  export default init;
+
+  /**
+   * Initialize the @tableland/sqlparser module.
+   * This is the synchronous counter-part to `init`, and is included primarily for testing purposes.
+   *
+   * @param bytes The input `bytes` must be a `BufferSource` to specify a local WASM file.
+   * @return A WASM `Exports` object.
+   */
+  export function initSync(bytes: BufferSource): WebAssembly.Exports;
+
+  /**
+   * The WASM `Exports` object cache.
+   */
+  export const __wasm: WebAssembly.Exports | undefined;
+}
+
+declare namespace sqlparser {
+  /**
+     * Validate and normalize a string containing (possibly multiple) SQL statement(s).
+
+     * @param statement A string containing SQL statement(s).
+     * @return A `Promise` that resolves to an array or normalized SQL statements.
+     */
+  export function normalize(statement: string): Promise<string[]>;
+}
